backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
pillow==10.4.0
django-cors-headers==4.4.0
boto3==1.34.154
django-storages==1.14.4
django-redis==5.4.0